requirements.txt
asgiref==3.9.1
Django==5.2.5
flake8==7.3.0
mccabe==0.7.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
sqlparse==0.5.3
---------------------- --------
asgiref                3.9.1
certifi                2025.8.3
charset-normalizer     3.4.3
Django                 5.2.5
django-extensions      4.1
django-jet-reboot      1.3.10
django-nested-admin    4.1.3
flake8                 7.3.0
greenlet               3.2.4
idna                   3.10
iniconfig              2.1.0
mccabe                 0.7.0
packaging              25.0
pillow                 11.3.0
pip                    25.0
playwright             1.54.0
pluggy                 1.6.0
psycopg2-binary        2.9.10
pycodestyle            2.14.0
pyee                   13.0.0
pyflakes               3.4.0
Pygments               2.19.2
pygraphviz             1.14
pytest                 8.4.1
pytest-django          4.11.1
python-monkey-business 1.1.0
requests               2.32.4
sqlparse               0.5.3
typing_extensions      4.14.1
urllib3                2.5.0